Reorganised files + Removed unnecessary files
