--- instafood_frontend/src/functions/displayImage.js
@@ -1,9 +1,0 @@
-function displayImage (image) {
-  return (
-    <div>
-        <img src={image} alt="post" />
-    </div>
-  )
-}
-
-export default displayImage;
