--- instafood_frontend/src/pages/Dashboard.css
@@ -1,105 +1,0 @@
-/*body {
-  background-image: url('../images/login_background.jpg');
-  background-size: cover;
-  background-position: center;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-}*/
-
-.container {
-  max-width: 800px;
-  margin: 0 auto;
-  padding: 40px;
-  background-color: rgba(255, 255, 255, 0.9);
-  border-radius: 10px;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
-}
-
-.welcome-message {
-  font-size: 32px;
-  color: #666;
-  margin-bottom: 40px;
-}
-
-.search-bar {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 20px;
-  background-color: #f2f2f2;
-  border-radius: 5px;
-}
-
-.search-bar p {
-  margin-bottom: 10px;
-}
-
-.search-bar input[type="text"] {
-  margin-bottom: 10px;
-  padding: 5px;
-  width: 200px;
-  border: 1px solid #ccc;
-  border-radius: 3px;
-}
-
-/* Custom Dropdown Styling */
-.search-bar select {
-margin-bottom: 10px;
-padding: 5px;
-width: 200px;
-border: 1px solid #ccc;
-border-radius: 3px;
-appearance: none;
-background-color: #fff;
-background-image: linear-gradient(45deg, transparent 50%, #777 50%),
-  linear-gradient(135deg, #777 50%, transparent 50%);
-background-position: right 10px top 50%, right 8px center;
-background-size: 5px 5px, 8px 8px;
-background-repeat: no-repeat;
-}
-
-.search-bar select::-ms-expand {
-display: none;
-}
-
-/* Hover & Focus Styles */
-.search-bar select:hover {
-border-color: #999;
-}
-
-.search-bar select:focus {
-border-color: #555;
-outline: none;
-}
-
-/* Firefox Focus Style Fix */
-@-moz-document url-prefix() {
-.search-bar select {
-  color: #000;
-}
-}
-
-
-.search-bar div {
-  margin-top: 10px;
-}
-
-.search-bar button {
-  padding: 8px 16px;
-  background-color: #4caf50;
-  color: #fff;
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-}
-
-.search-bar button:hover {
-  background-color: #45a049;
-}
-
-.search-bar button:focus {
-  outline: none;
-}
